chore: move and rename tooltip
